Refactor translation service into modules and extract middleware
