+++ backend/content-service/src/middleware/security.ts
@@ -1,0 +1,35 @@
+import express from 'express';
+import helmet from 'helmet';
+import rateLimit from 'express-rate-limit';
+
+export function applySecurityMiddleware(app: express.Application): void {
+    app.use(helmet({
+        contentSecurityPolicy: {
+            directives: {
+                defaultSrc: ["'self'"],
+                styleSrc: ["'self'", "'unsafe-inline'"],
+                scriptSrc: ["'self'"],
+                imgSrc: ["'self'", 'data:', 'https:'],
+                connectSrc: ["'self'"],
+                fontSrc: ["'self'"],
+                objectSrc: ["'none'"],
+                mediaSrc: ["'self'"],
+                frameSrc: ["'none'"],
+            },
+        },
+        hsts: {
+            maxAge: 31536000,
+            includeSubDomains: true,
+            preload: true,
+        },
+    }));
+
+    const generalLimiter = rateLimit({
+        windowMs: 15 * 60 * 1000,
+        max: 300,
+        message: { error: 'Too many requests, please try again later.' },
+        standardHeaders: true,
+        legacyHeaders: false,
+    });
+    app.use(generalLimiter);
+}
